sitecheck/requirements.txt
asgiref==3.3.0
certifi==2020.12.5
chardet==3.0.4
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.10
lxml==4.6.0
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.0
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
urllib3==1.26.2
xlrd==1.2.0
